+++ src/SMAPI.Web/Framework/Clients/Pastebin/PasteInfo.cs
@@ -11,12 +11,6 @@ namespace StardewModdingAPI.Web.Framework.Clients.Pastebin
         /// <summary>The fetched paste content (if <see cref="Success"/> is <c>true</c>).</summary>
         public string Content { get; set; }
 
-        /// <summary>When the file will no longer be available.</summary>
-        public DateTime? Expiry { get; set; }
-
-        /// <summary>The error message if saving succeeded, but a non-blocking issue was encountered.</summary>
-        public string Warning { get; set; }
-
         /// <summary>The error message if saving failed.</summary>
         public string Error { get; set; }
     }